Techno-Commercial-analysis-of-Electric-Vehicle
The project is a Techno-Commercial analysis of the Tata Nexon EV which is an electric vehicle and runs on batteries. Uncovered in the year 2019. The Tata Nexon EV is Tata Motors' second creation
electric vehicle and depends on the current Tata Nexon. In this report, I have examined the techno-commercial highlights of Tata Nexon EV, what it means for our general public and how it tends to be advantageous. Strategy for examination incorporates pattern, vertical and even correlations, diagrams. It has a single electric motor driving the front wheels 95 kW and 245 Nm—enough to take the 4.3-meter-long hybrid from 0–60 kph in 4.6 seconds. The battery size is 30.2kWh. Tata Nexon EV's high energy thickness Lithium-battery pack is viable with CCS 2 quick charging standard and supports home and office accusing of any 15A attachment point.